- solved issue #307 - solved issue #304 - solved issue #308 - solved issue #309 - solved issue #291
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@1162 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ sql/hive-0.0.3.sql
@@ -47,7 +47,7 @@ CREATE TABLE IF NOT EXISTS `files` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `filename` varchar(255) DEFAULT NULL,
   `path` varchar(255) NOT NULL,
-  `size` int(11) NOT NULL DEFAULT '0',
+  `size` bigint(20) NOT NULL DEFAULT '0',
   `stream_count` int(2) NOT NULL DEFAULT '0',
   `container_type` varchar(255) NOT NULL,
   `title` varchar(255) NOT NULL,
